% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch an interactive learnr tutorial}
\usage{
launch(tute)
}
\arguments{
\item{tute}{A learnr tutorial}
}
\description{
Launch an interactive learnr tutorial
}
\examples{
launch("dags")
}
