% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Launch an interactive Shiny app}
\usage{
run(app)
}
\arguments{
\item{app}{A Shiny app. Valid apps are: "dags-1", "backdoor-paths", "collider-bias", "matching-1", "matching-2", "mlm-1", "missing-data-1"}
}
\description{
Launch an interactive Shiny app
}
\examples{
run("missing-data-1")
}
